+++ packages/picasso-lab/src/EmptyStatePage/styles.ts
@@ -1,0 +1,9 @@
+import { createStyles } from '@material-ui/core/styles'
+
+export default () =>
+  createStyles({
+    image: {
+      width: '8rem',
+      height: '8rem'
+    }
+  })