+++ packages/picasso-lab/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.35.0](https://github.com/toptal/picasso/compare/@toptal/picasso-lab@2.34.2...@toptal/picasso-lab@2.35.0) (2020-10-29)
+
+
+### Features
+
+* accept node instead only text ([#1647](https://github.com/toptal/picasso/issues/1647)) ([48c830b](https://github.com/toptal/picasso/commit/48c830b94f23915b2946df00c50fd70fb722d937))
+
+
+
+
+
+## [2.34.2](https://github.com/toptal/picasso/compare/@toptal/picasso-lab@2.34.1...@toptal/picasso-lab@2.34.2) (2020-10-27)
+
+**Note:** Version bump only for package @toptal/picasso-lab
+
+
+
+
+
+## [2.34.1](https://github.com/toptal/picasso/compare/@toptal/picasso-lab@2.34.0...@toptal/picasso-lab@2.34.1) (2020-10-21)
+
+**Note:** Version bump only for package @toptal/picasso-lab
+
+
+
+
+
+# [2.34.0](https://github.com/toptal/picasso/compare/@toptal/picasso-lab@2.33.0...@toptal/picasso-lab@2.34.0) (2020-10-20)
+
+
+### Features
+
+* **TypographyOverflow:**  add custom tooltipContent ([#1632](https://github.com/toptal/picasso/issues/1632)) ([c06f9f7](https://github.com/toptal/picasso/commit/c06f9f7855186dc13bb6addd9f087337c571e9a8))
+
+
+
+
+
+# [2.33.0](https://github.com/toptal/picasso/compare/@toptal/picasso-lab@2.32.3...@toptal/picasso-lab@2.33.0) (2020-10-16)
+
+
+### Features
+
+* **EmptyState:** add Page and List empty components  ([#1556](https://github.com/toptal/picasso/issues/1556)) ([76e76b3](https://github.com/toptal/picasso/commit/76e76b3f1682ffe7a79553d5a59e85cdab3dc2ef))
+
+
+
+
+
 ## [2.32.3](https://github.com/toptal/picasso/compare/@toptal/picasso-lab@2.32.2...@toptal/picasso-lab@2.32.3) (2020-09-23)